third_party/WebKit/LayoutTests/editing/inserting/prevent-block-nesting-01.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    '<div contenteditable>|<br></div>',
    'insertHTML abc<span><div><br></div></span><div>def</div>',
    '<div contenteditable>abc<div><br></div><div>def|</div></div>'),
    'insertHTML prevents block nesting');
</script>
